Minor code refactor. Starting to implement authentication
