+++ backend/models/user.js
@@ -1,0 +1,20 @@
+const mongoose = require('mongoose');
+
+const { Schema } = mongoose;
+
+const UserSchema = new Schema({
+  name: {
+    type: String,
+    required: true
+  },
+  username: {
+    type: String,
+    required: true
+  },
+  password: {
+    type: String,
+    required: true
+  }
+});
+
+module.exports = mongoose.model('User', UserSchema);
