+++ client/src/actions/postsActions.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import * as types from './actionTypes';
 
 export const getPosts = () => dispatch =>
-  axios.get('/api/posts').then((res) => {
+  axios.get('/posts').then((res) => {
     dispatch({
       type: types.GET_POSTS,
       payload: res.data
@@ -10,7 +10,7 @@ export const getPosts = () => dispatch =>
   });
 
 export const createPost = text => dispatch =>
-  axios.post('/api/posts', { text, author: 'unidentified' }).then((res) => {
+  axios.post('/posts', { text, author: 'unidentified' }).then((res) => {
     dispatch({
       type: types.CREATE_POST,
       payload: res.data
@@ -18,7 +18,7 @@ export const createPost = text => dispatch =>
   });
 
 export const updatePost = (id, text, author) => dispatch =>
-  axios.patch(`/api/posts/${id}`, { id, text, author }).then((res) => {
+  axios.patch(`/posts/${id}`, { id, text, author }).then((res) => {
     dispatch({
       type: types.UPDATE_POST,
       id,
@@ -28,7 +28,7 @@ export const updatePost = (id, text, author) => dispatch =>
   });
 
 export const deletePost = id => dispatch =>
-  axios.delete(`/api/posts/${id}`).then((res) => {
+  axios.delete(`/posts/${id}`).then((res) => {
     dispatch({
       type: types.DELETE_POST,
       id
