+++ package.json
@@ -22,12 +22,17 @@
   },
   "homepage": "https://github.com/jm-shi/MERN-Social-Network#readme",
   "dependencies": {
+    "bcrypt": "^3.0.0",
     "body-parser": "^1.18.3",
     "concurrently": "^4.0.1",
     "express": "^4.16.3",
+    "jsonwebtoken": "^8.3.0",
     "mongodb": "^3.1.4",
-    "mongoose": "^5.2.12",
-    "morgan": "^1.9.0"
+    "mongoose": "^5.2.13",
+    "morgan": "^1.9.0",
+    "passport": "^0.4.0",
+    "passport-jwt": "^4.0.0",
+    "validator": "^10.7.1"
   },
   "devDependencies": {
     "babel-eslint": "^9.0.0",
